server/requirements.txt
fastapi==0.73.0
uvicorn==0.17.6
RPi.GPIO==0.7.0